Avoid overhead of allocating and freeing GCHandles by caching longer
In benchmarks the MarshalNative::GCHandleInternalAlloc method shows up noticably.  This is because a GC handle
is created and destroyed on each Write request.   Change it so that we keep the GCHandle for the lifetime of the request structure
instead.

There was already a 'Pin' (which is a bad name) that causes GCHandle to point at the current object,   Change the WriteRequest
to use this API, and then change that API to be lazy, and to only free the Handle when the object is destroyed (ReleaseHandle).

This change has a positive effect immediately, but most of its value comes when we optimize the .NET Runtime to make
updating a GCHandle target faster (target update can be fast, allocation will always be slower).